Add PostgreSQL icon and renamed a Sass function
This commit includes an addition of 'PostgreSQL' icon in 'getIcons.ts'.
Also, the function 'getNumbersFromSassProperty' is renamed to 'getNumbersFromCssProperty' for reflecting the correct type of files it processes.

This function change is necessary because it aids in better reflection of the file types the function deals with, and the Icon addition would facilitate the users expecting a PostgreSQL icon. Addition of icons generally aims for better user experience.