Update npm-check-updates version from 16.10.17 to 16.10.18
The version of the npm-check-updates dependency has been updated from 16.10.17 to 16.10.18. This update was carried out as per the recommendations from the npm community to keep the dependencies updated to their latest versions as they may include bug fixes, security updates, and performance improvements. This modification aids in maintaining the robustness of the project.