Update PostStubs.svelte to dynamically set the displayedPostsLength based on the URL hash and selected categories/tags.
Previously, the isHashInUrl variable was used to determine if the URL contained a hash (#), but it was unused. This variable has now been removed.

The code has been updated to use the window.location.href.includes("#") condition directly to check if the URL contains a hash. If it does, the displayedPostsLength is set to MAX_POSTS_LENGTH.
If the URL does not contain a hash but there are selected categories or tags, the displayedPostsLength is also set to MAX_POSTS_LENGTH.

These modifications ensure that the correct number of posts are displayed based on the URL and selected categories or tags.