Update displayed post length logic in PostStubs component
The logic for determining the displayed post length in the PostStubs component has been updated. Previously, the length was only set to `MAX_POSTS_LENGTH` if the window included a `#` in the URL. However, now it will also be set to `MAX_POSTS_LENGTH` if either `$showCategories` or `$showTags` have a size greater than 0. This change ensures that the correct number of posts are displayed based on the selected categories or tags.