feat: Add experience as a Technical Pedagogical Accompanist at Epitech Digital
Added a new file "epitech.md" in the "content" directory, which contains the experience details of working as a Technical Pedagogical Accompanist at Epitech Digital School Lyon. The file includes information such as the job title, responsibilities, skills, website link, and other relevant details. Also updated the "influactive.md" file in the "content" directory to mark it as not ongoing. In addition, made a textual change to the "src/routes/item/[slug]/+page.svelte" file to change the label from "Website" to "Site web".