feat: Add eslint-plugin-svelte to package-lock.json
This commit adds the `eslint-plugin-svelte` package to `package-lock.json`.

Upgrade script in package.json modified to exclude `universal-cookie`.

The `files` list in `wireit` configuration in package.json has been updated to include all files in the `src` directory and its subdirectories.

In the `getPosts` script, the `files` list has been modified to include all files in the `content` directory and its subdirectories.